+++ src/IIIFPresentation/Models/API/General/DeleteCollectionType.cs
@@ -6,14 +6,3 @@ public enum DeleteCollectionType
     CollectionNotEmpty = 2,
     Unknown = 3
 }
-
-public enum ModifyCollectionType
-{
-    DuplicateSlugValue = 1,
-    UnknownDatabaseSaveError = 2,
-    ETagNotAllowed = 3,
-    ParentCollectionNotFound = 4,
-    ETagNotMatched = 5,
-    PossibleCircularReference = 6,
-    CannotGenerateUniqueId = 7,
-}
